requirements.txt
astroid==1.5.3
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.4
PySocks==1.6.7
six==1.11.0
wrapt==1.10.11
